Start of Pink blog
A blogposts.py template has been created to give me content to work with.

htmldominate.py was to test the dominate library for page creation.

htmljinja.py is currently the main creator for pages.

style.css holds the current styling attributes for the page.

simple-variables.html is my current building page for html.