Created more pages for the layout
Worked on diving the blog draft content into different SQL tables, and gone over them manually to make sure the content matches the ID numbers.

The general css document has been tweaked to include new elements, and with an added effort of avoiding code duplication.

The app.py page has been created, and gained a lot of elements, to make sure the additional pages work that will be running in the background.

For the blogarchive, I want to do some alterations to the blog posts to give the site a better display that is not just a copy-paste from the main page.
For now, they look very similar.

The create blog post, edit post, simple variable and single post all look pretty similar so far.

I added an edit icon to the single post page, to make a slight alteration to the page and not have it only dependant on text. Might go and add more icons later on.