SQL adjustments, tags page
I'm having some issues linking my blog.db file link up with the page, so a temporary solution I've done is to be explicit in the location for the db file. This needs to be fixed.

Ive created a tag page, which offers a search option. It needs some more work so that clicking on a tag shows the blog posts that use that tag.

Next step is then to allow users to click on a tag on the main page, and let them see all the posts that share the tag clicked.

I plan on going right back to working on this, but after my mac crashed, I'm commiting these files to the main so that I know things are saved and available.