Tests, tags, links and comments
The links on all pages should now work, and link to the desired pages.

README.md has become a very handy overview of tasks, and they are slowly going down.

test_app.py and test_integration.py are two files that run at the same time, and neither is returning errors, but one is returing a few warnings.

app.py has been updated to look for, and potentially create, a blog.db in the instance folder. Previously I used an explicit path to locate the file. I didn't expect SQLAlchemy to automatically create a file in the instance folder, but now I'm making sure that file has the tables with content I want instead of making an empty file.

All tags are now capitalized automatically, meaning mindfullness and Mindfullness will be linked up.

The blog.sql has gotten a new table, for now it is empty, I might want to add some comments there to showcase what that would look like.

Tags, forms and testers have been a big focus for this round of edits.