CSS + Tags search function
Fixed the search_tags() function so when an unused element is searched for, the query element is returned to the tags.html template, to show what you searched for.

More CSS has been added. I've kept it to one page as the CSS isn't the main focus, and made sure to go through other pages, making matching elements share class names to match with the CSS code already sat up.

blogarchive has gotten a small change. My app.py, used this page as a defualt for showing posts with the same tags, so now the title can be swapped out depending on what URL you're using.

single-post-page has had a LOT of code moved into a new DIV, that way I could add a border around the area with CSS.

tags.html have gotten linked up with CSS, and the result for not finding a post with the searched tag, now has been styled.