Comment explainers, README.md update, CSS cleanup
First branch creation, I should have remembered to use the feature more.

Comments added over most elements in app.py to make it easy to look for relevant function in case something is wrong.

Deleted the old blog.db. app.py can now create a new one when it is run.

README.md update, added explination for the blog.db document, as I struggled with this issue earlier.

Removed an element in style.css that was not a correct function.